agents/freedom_agent/requirements.txt
# Pinned to a known-stable range. livekit-agents 0.12.x+ supports AgentSession API.
livekit-agents[openai]>=0.12,<1.0
python-dotenv>=1.0
